Implemented the Update function

+++ src/main/java/com/hacknc/FarmMap.java
@@ -5,7 +5,7 @@ import java.io.IOException;
 
 public interface FarmMap {
     //Updates the farmMap based on deltaTime
-    public void update(long deltaTime) throws IOException;
+    public void update(double deltaTime) throws IOException;
 
     //Returns the panel that is the map.
     public JPanel getMap();
